Feat- Update more accurate score count and add in ScoreFile value profiles

+++ backend/src/main/java/com/eris/gitlabanalyzer/controller/ScoreController.java
@@ -20,12 +20,12 @@ public class ScoreController {
     }
 
     @GetMapping(path ="/projects/{projectId}/merge_request/{merge_request_id}/diff/score")
-    public int getMergeDiffScore (@PathVariable("projectId") Long projectId,
+    public double getMergeDiffScore (@PathVariable("projectId") Long projectId,
                              @PathVariable("merge_request_id") Long merge_request_id){
         return scoreService.getMergeDiffScore(projectId, merge_request_id);
     }
     @GetMapping(path ="/projects/{projectId}/merge_requests/score")
-    public int getTotalMergeDiffScore (@PathVariable("projectId") Long projectId,
+    public double getTotalMergeDiffScore (@PathVariable("projectId") Long projectId,
                                        @RequestParam("startDateTime")
                                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime startDateTime,
                                        @RequestParam("endDateTime")
@@ -34,12 +34,12 @@ public class ScoreController {
     }
 
     @GetMapping(path ="/projects/{projectId}/commit/{commitId}/diff/score")
-    public int getCommitDiffScore (@PathVariable("projectId") Long projectId,
+    public double getCommitDiffScore (@PathVariable("projectId") Long projectId,
                              @PathVariable("commitId") Long commitId){
         return scoreService.getCommitDiffScore(projectId, commitId);
     }
     @GetMapping(path ="/projects/{projectId}/commits/score")
-    public int getTotalCommitDiffScore (@PathVariable("projectId") Long projectId,
+    public double getTotalCommitDiffScore (@PathVariable("projectId") Long projectId,
                                        @RequestParam("startDateTime")
                                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime startDateTime,
                                        @RequestParam("endDateTime")
